+++ src/js/app/screen/status.js
@@ -71,6 +71,7 @@ app.screen.status = (() => {
     const {x, y, z} = engine.position.getVector()
 
     const coordinates = {x, y},
+      terrain = content.terrain.current(),
       time = content.time.value(),
       velocity = engine.position.getVelocity(),
       yaw = engine.position.getEuler().yaw
@@ -78,6 +79,7 @@ app.screen.status = (() => {
     root.querySelector('.a-status--metric-altitude').innerHTML = app.utility.format.number(z)
     root.querySelector('.a-status--metric-coordinates').innerHTML = app.utility.format.coordinates(coordinates)
     root.querySelector('.a-status--metric-heading').innerHTML = app.utility.format.angle(yaw)
+    root.querySelector('.a-status--metric-height').innerHTML = app.utility.format.number(Math.max(z - terrain, 0))
     root.querySelector('.a-status--metric-time').innerHTML = app.utility.format.time(time)
     root.querySelector('.a-status--metric-velocity').innerHTML = app.utility.format.velocity(velocity)
   }